[FIXED] Deterministic clustered dedupe (#6415)
Issue description:
- Publish a message with a `Nats-Msg-Id`.
- Leader election through restarts/networking/etc.
- Due to leader election the proposal queue is drained and the message
is not stored/proposed.
- Any retries of the original message will fail until the de-dupe window
clears the original message, even though it was not stored in the
stream.

This issue is due to staging a zero sequence in the de-dupe map. An easy
solution seems to clear all zero sequence entries from the de-dupe map
upon stepping down, but that has correctness issues if any of those
messages did get proposed.

This PR removes the staging of a zero sequence and ensures all replicas
can deterministically do de-duping themselves. And duplicate messages
are only blocked at the cluster-level if we know what the sequence is.

Signed-off-by: Maurice van Veen <github@mauricevanveen.com>